+++ package.json
@@ -13,7 +13,7 @@
     "@headlessui/react": "^2.2.4",
     "@heroicons/react": "^2.2.0",
     "@vitejs/plugin-react": "^4.3.4",
-    "framer-motion": "^12.12.1",
+    "framer-motion": "^12.14.0",
     "i18next": "^24.2.0",
     "i18next-browser-languagedetector": "^8.0.2",
     "i18next-http-backend": "^3.0.1",
