+++ src/pages/Layout/Layout.jsx
@@ -1,18 +1,7 @@
 import { Suspense } from "react";
 import { Outlet } from "react-router-dom";
 import Header from "../../components/organisms/Header";
-import Footer from "../../footer/Footer";
-import Forms from "../../footer/Forms";
-import { Frontend } from "../directions/Frontend";
 
-import { Information } from "../directions/Information";
-import { AutoMechanic } from "../directions/AutoMechanic";
-
-import { Backend } from "../directions/Backend";
-import { MobileDevelopment } from "../directions/MobileDevelopment";
-import { TD } from "../directions/TD";
-import { SisAdm } from "../directions/SisAdm";
-import { MechRob } from "../directions/MechRob";
 
 
 const Layout = () => {
